requirements.txt
openai==0.28
python-dotenv
gradio
chromadb
langchain
langchain-community
langchain-core
tiktoken
datetime
pypdf